Do 07 flex exercises
Remove underlining on anchor. Set header and footer margin to 0px. Center text in header and footer using flex alig and justify. Add flex display to body, justify space between, and flex-direction column. Set the deck class to flex, flex warp: warp, and add padding. Add a div to contain the side bar and the deck: class = bardeck. Set bardeck with padding and margin to 0. Set flex shrink of sidebar to 0. Set bardeck flex-grow to 1 to fill all the space.